Cleanup handling if user wants to use karma or jest. Added some documentation
