+++ lib/ui.js
@@ -10,7 +10,6 @@ function setupUI(config){
     port,
     webpackConfig,
     karmaConfig,
-    jestConfig,
     proxy: {
       endpoints,
       server
@@ -22,7 +21,6 @@ function setupUI(config){
       webpackDevMiddleware,
       webpackHotMiddleware,
       karma,
-      jest,
     }
   } = config;
 
@@ -60,20 +58,13 @@ function setupUI(config){
     handler: setupProxyLogs(server)
   }];
 
-    if (jest) {
-      childrenConfig.push({
-        file: '../client/jestServer.js',
-        args: [`--config=${jestConfig}`, `--module=${jest}`],
-        logBoxConfig: { top: 0, right: 0, height: '50%', label: 'Tests'}
-      });
-    } else if (karma) {
-      childrenConfig.push({
-        file: '../client/karmaServer.js',
-        args: [`--config=${karmaConfig}`, `--module=${karma}`],
-        logBoxConfig: { top: 0, right: 0, height: '50%', label: 'Tests'}
-      });
-    }
-    var children = childrenConfig.reduce((children, config) => {
+  childrenConfig.push({
+    file: '../client/testServer.js',
+    args: karma ? [`--config=${karmaConfig}`, `--module=${karma}`] : [],
+    logBoxConfig: { top: 0, right: 0, height: '50%', label: 'Tests'}
+  });
+
+  var children = childrenConfig.reduce((children, config) => {
     var child = child_process.fork(path.join(__dirname, config.file), config.args, {silent:true});
     var box = buildLogBox(config.logBoxConfig, child);
     screen.append(box);
